refactor: Organize folder structure (MVC)
